requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.1
pyjokes==0.6.0
pywhatkit==5.4
wikipedia==1.4.0
